requirements.txt
crewai==0.41.1
crewai-tools==0.4.26
langchain-core==0.2.43
langchain-google-genai==1.0.10
google-generativeai==0.7.2
pydantic==2.8.2
python-dotenv==1.0.1
streamlit==1.36.0